refactor: deduplicate http_types.rs and chat_route.rs → re-exports from astra-server-types
Replace 2,243 lines of duplicate type definitions in runtime's server/http_types.rs
and server/chat_route.rs with re-exports from astra-server-types where the canonical
definitions already live.

Changes:
- server/http_types.rs (1,603 LOC → 14 LOC): explicit re-exports of 46 types/fns
- server/chat_route.rs (638 LOC → 1 LOC): re-export ChatRouteResponse + classify_chat_route
- Move 73 http_type tests to astra-server-types/src/http_type_tests.rs
- Remove unused imports: regex::RegexBuilder, RunMutationRecord from lib.rs

Net: -2,224 LOC from runtime crate, 0 behavioral changes.
